+++ app/Http/Controllers/ProductsController.php
@@ -7,12 +7,20 @@ use Illuminate\Support\Facades\Validator;
 use Purifier;
 use Response;
 use App\Product;
+use Auth;
 
 class ProductsController extends Controller
 {
     public function index()
     {
       $product = Product::all();
+
+      $user = Auth::user();
+      if($user->roleID != 1)
+
+      {
+        return Response::json(["error" => "Not Allowed"]);
+      }
 
       return Response::json($product);
     }
@@ -33,6 +41,20 @@ class ProductsController extends Controller
     if($validator->fails())
     {
       return Response::json(["error" => "You need to fill out all fields."]);
+    }
+
+    $user = Auth::user();
+    if($user->roleID != 1)
+
+    {
+      return Response::json(["error" => "Not Allowed"]);
+    }
+
+    $category = Category::find($request->input("categoryID"));
+    if(empty($category))
+
+    {
+      return Response::json(["error" => "Category not found!"]);
     }
 
     $product = new Product;
@@ -70,6 +92,13 @@ class ProductsController extends Controller
       return Response::json(["error" => "You need to fill out all fields."]);
     }
 
+    $user = Auth::user();
+    if($user->roleID != 1)
+
+    {
+      return Response::json(["error" => "Not Allowed"]);
+    }
+
     $product = Product::find($id);
 
     $product->name = $request->input('name');
@@ -92,12 +121,26 @@ class ProductsController extends Controller
     {
       $product = Product::find($id);
 
+      $user = Auth::user();
+      if($user->roleID != 1)
+
+      {
+        return Response::json(["error" => "Not Allowed"]);
+      }
+
       return Response::json($product);
     }
 
   public function destroy($id)
     {
       $product = Product::find($id);
+
+      $user = Auth::user();
+      if($user->roleID != 1)
+
+      {
+        return Response::json(["error" => "Not Allowed"]);
+      }
 
       $product->delete();
